pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smoke-diff-cli"
version = "0.1.0"
description = "Generate and run smoke tests from API definitions"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "pytest>=7.4.0",
    "rich>=13.0.0",
    "deepdiff>=6.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
smoke-diff = "smoke_diff_cli.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["smoke_diff_cli*"]
